+++ app/models/project.rb
@@ -3,10 +3,13 @@ class Project < ActiveRecord::Base
   # Concerns
   include Deletable
 
+  has_secure_password
+
   # Model Validation
   validates_presence_of :name, :user_id
   validates_uniqueness_of :slug, scope: [ :deleted ], allow_blank: true
   validates_format_of :slug, with: /\A[a-z][a-z0-9\-]*\Z/, allow_blank: true
+  validates_uniqueness_of :username, :number, allow_blank: true
 
   # Model Relationships
   belongs_to :user