+++ db/migrate/20150616225036_add_username_and_password_to_projects.rb
@@ -1,0 +1,9 @@
+class AddUsernameAndPasswordToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :username, :string
+    add_column :projects, :password_plain, :string
+    add_column :projects, :password_digest, :string
+
+    add_index :projects, :username, unique: true
+  end
+end