+++ db/migrate/20150616231430_add_client_name_to_projects.rb
@@ -1,0 +1,5 @@
+class AddClientNameToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :client_name, :string
+  end
+end