+++ db/migrate/20150616231700_add_number_to_projects.rb
@@ -1,0 +1,5 @@
+class AddNumberToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :number, :string
+  end
+end